Clean up mirage scenarios
